[TOSA] Rescale output_zp fix
Patch corrects output_zp in case of usigned output by removing
sign-extended part.
Also corrects check in verifyZeroPoint for int16 type

+++ mlir/lib/Dialect/Tosa/IR/TosaOps.cpp
@@ -1851,7 +1851,8 @@ static LogicalResult verifyZeroPoint(tosa::RescaleOp op, Value zpVal,
       return op.emitOpError()
              << "expect " << tensorName << "_zp of 0, got " << zp;
     }
-    if (zpElemType.isInteger(16) && tensorUnsigned && zp != 32768) {
+    if (zpElemType.isInteger(16) && tensorUnsigned &&
+        zp != static_cast<int16_t>(32768)) {
       return op.emitOpError() << "expect " << tensorName
                               << "_zp of 0 or 32768 for unsigned int16 "
                               << tensorName << ", got " << zp;
